+++ gdpr/src/main/java/com/kairos/controller/master_data/processing_activity_masterdata/TransferMethodController.java
@@ -50,7 +50,7 @@ public class TransferMethodController {
 
     @ApiOperation("get transfer Method by id")
     @GetMapping("/transfer_method/{transferMethodId}")
-    public ResponseEntity<Object> getTransferMethod(@PathVariable Long countryId, @PathVariable BigInteger transferMethodId) {
+    public ResponseEntity<Object> getTransferMethod(@PathVariable Long countryId, @PathVariable Long transferMethodId) {
         return ResponseHandler.generateResponse(HttpStatus.OK, true, transferMethodDestinationService.getTransferMethod(countryId, transferMethodId));
     }
 
@@ -63,21 +63,21 @@ public class TransferMethodController {
 
     @ApiOperation("delete transfer Method by id")
     @DeleteMapping("/transfer_method/{transferMethodId}")
-    public ResponseEntity<Object> deleteTransferMethod(@PathVariable Long countryId, @PathVariable BigInteger transferMethodId) {
+    public ResponseEntity<Object> deleteTransferMethod(@PathVariable Long countryId, @PathVariable Long transferMethodId) {
         return ResponseHandler.generateResponse(HttpStatus.OK, true, transferMethodDestinationService.deleteTransferMethod(countryId, transferMethodId));
     }
 
 
     @ApiOperation("update transfer Method by id")
     @PutMapping("/transfer_method/{transferMethodId}")
-    public ResponseEntity<Object> updateTransferMethod(@PathVariable Long countryId, @PathVariable BigInteger transferMethodId, @Valid @RequestBody TransferMethodDTO transferMethod) {
+    public ResponseEntity<Object> updateTransferMethod(@PathVariable Long countryId, @PathVariable Long transferMethodId, @Valid @RequestBody TransferMethodDTO transferMethod) {
         return ResponseHandler.generateResponse(HttpStatus.OK, true, transferMethodDestinationService.updateTransferMethod(countryId, transferMethodId, transferMethod));
     }
 
 
     @ApiOperation("update Suggested status of Transfer methods")
     @PutMapping("/transfer_method")
-    public ResponseEntity<Object> updateSuggestedStatusOfTransferMethods(@PathVariable Long countryId, @RequestBody Set<BigInteger> transferMethodIds, @RequestParam(required = true) SuggestedDataStatus suggestedDataStatus) {
+    public ResponseEntity<Object> updateSuggestedStatusOfTransferMethods(@PathVariable Long countryId, @RequestBody Set<Long> transferMethodIds, @RequestParam(required = true) SuggestedDataStatus suggestedDataStatus) {
         if (CollectionUtils.isEmpty(transferMethodIds)) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "Transfer Method is Not Selected");
         } else if (!Optional.ofNullable(suggestedDataStatus).isPresent()) {
